+++ src/__playground__/worker.ts
@@ -1,16 +1,7 @@
-import { ParentThread, define } from 'nanolith';
-import fs from 'fs';
+import { define } from 'nanolith';
 
 export const worker = await define({
     __initializeService() {
-        ParentThread.onStream((stream) => {
-            const write = fs.createWriteStream('./foo.txt');
-
-            stream.pipe(write);
-
-            write.on('finish', () => {
-                ParentThread.sendMessage('ready_to_close');
-            });
-        });
+        console.log('Initialized');
     },
 });